% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_other_median.R
\name{my_other_median}
\alias{my_other_median}
\title{My Other median}
\usage{
my_other_median(x, na.rm = TRUE)
}
\arguments{
\item{x}{Vector of Numeric values}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}
}
\value{
Median of vector x
}
\description{
My Other median
}
\examples{
my_other_median(1:12)
}
