% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfusen.demo-package.R
\docType{package}
\name{Rfusen.demo-package}
\alias{Rfusen.demo}
\alias{Rfusen.demo-package}
\title{Rfusen.demo: Build A Package From Rmarkdown File}
\description{
Use Rmarkdown First method to build your package. Start your package with documentation. Everything can be set from a Rmarkdown file in your project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/Rfusen.demo}
  \item Report bugs at \url{https://github.com/jimbrig/Rfusen.demo/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
