% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_other_median.R
\name{sub_median}
\alias{sub_median}
\title{Core of the median not exported}
\usage{
sub_median(x, na.rm = TRUE)
}
\arguments{
\item{x}{Vector of Numeric values}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}
}
\description{
Core of the median not exported
}
